Add support for nginx
Nginx doesn't have the htaccess mechanism. Redirection rules need to be
specified in the nginx configuration. This is less flexible. We create a
nginx configuration file with the redirect rules and this file should be
included in the nginx configuration for the web server with:
include <path-to-redirect-file>

Another disadvantage is that the nginx server needs a restart on a
configuration change.